+++ Evaluator.cs
@@ -123,6 +123,16 @@ public abstract class Evaluator
         string contextName, string? contextStateName = null);
 
     /// <summary>
+    /// Provides the operations of the specified behavior type that should 
+    /// be invoked when the behavior is being torn down.
+    /// </summary>
+    /// <param name="behaviorType">The type of the behavior whose operations 
+    /// are to be provided.</param>
+    /// <returns>The behavior type's operations that should be invoked 
+    /// when the behavior is being torn down.</returns>
+    public abstract MethodInfo[] GetOnTeardownOperations(Type behaviorType);
+
+    /// <summary>
     /// Specifies whether the provided type is a type of context known by this evaluator.
     /// </summary>
     /// <param name="type">The type to be checked.</param>
@@ -189,6 +199,11 @@ public class Evaluator<TContextAttribute, TBehaviorAttribute,
     /// </summary>
     private readonly Dictionary<Type, Dictionary<string, Dictionary<string, List<MethodInfo>>>>
         _onChangeOperations = new();
+
+    /// <summary>
+    /// All found operations that should be invoked upon teardown.
+    /// </summary>
+    private readonly Dictionary<Type, List<MethodInfo>> _onTeardownOperations = new();
 
 
     /// <inheritdoc/>
@@ -320,6 +335,23 @@ public class Evaluator<TContextAttribute, TBehaviorAttribute,
             return Array.Empty<MethodInfo>();
 
         return _onChangeOperations[behaviorType][contextName][stateName].ToArray();
+    }
+
+    /// <inheritdoc/>
+    public override MethodInfo[] GetOnTeardownOperations(Type behaviorType)
+    {
+        ValidateInitialization();
+
+        if (!_behaviorTypes.Contains(behaviorType))
+            throw new ArgumentException($"Teardown Operations cannot be retrieved " +
+                $"for type {behaviorType.FullName} since it is not a behavior " +
+                $"known to an evaluator with behaviors defined " +
+                $"by {typeof(TBehaviorAttribute).FullName}.");
+
+        if (!_onTeardownOperations.ContainsKey(behaviorType))
+            return Array.Empty<MethodInfo>();
+
+        return _onTeardownOperations[behaviorType].ToArray();
     }
 
     /// <inheritdoc/>
@@ -484,6 +516,7 @@ public class Evaluator<TContextAttribute, TBehaviorAttribute,
                 ValidateOperation(behaviorType, methods[c]);
 
                 CacheOnChangeOperations(behaviorType, methods[c]);
+                CacheOnTeardownOperations(behaviorType, methods[c]);
                 // TODO :: Support other types of operation caching here.
             }
     }
@@ -516,6 +549,26 @@ public class Evaluator<TContextAttribute, TBehaviorAttribute,
 
             _onChangeOperations[behaviorType][dn][sn].Add(operation);
         }
+    }
+
+    /// <summary>
+    /// Validates and caches the provided operation as an on teardown operation 
+    /// for each such declaration found to be associated with the operation.
+    /// </summary>
+    /// <param name="behaviorType">The type of behavior whose operation is being evaluated.</param>
+    /// <param name="operation">The operation being evaluated, validated, and cached 
+    /// for on teardown declarations.</param>
+    private void CacheOnTeardownOperations(Type behaviorType, MethodInfo operation)
+    {
+        IEnumerable<OnTeardownAttribute> attrs = operation
+            .GetCustomAttributes<OnTeardownAttribute>(true);
+        if (attrs.Count() == 0)
+            return;
+
+        if (!_onTeardownOperations.ContainsKey(behaviorType))
+            _onTeardownOperations.Add(behaviorType, new());
+
+        _onTeardownOperations[behaviorType].Add(operation);
     }